+++ src/parser.ts
@@ -427,6 +427,11 @@ function generateExtraMetadata(
   }
 
   Object.keys(node).forEach((key) => {
+    if (key === "inputParameters" && Array.isArray(node.inputParameters)) {
+      node.inputParameters.forEach((inputParameter: any) => {
+        inputParameter.insideParenthesis = true;
+      });
+    }
     if (typeof node[key] === "object") {
       if (Array.isArray(node)) {
         const keyInt = parseInt(key, 10);